LayoutTests/platform/ios-simulator/fast/events/touch/touch-before-pressing-spin-button-expected.txt

Test if a spin-button works correctly after touch events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL numberInput.value should be 2. Was 1.
PASS successfullyParsed is true

TEST COMPLETE
